Send meters read successfully or failed to queues

+++ MeterReading/Domain/Meter.cs
@@ -29,7 +29,7 @@ public class Meter
     public required string UserId { get; set; }
 
     [BsonElement("lastReading")]
-    public double LastReading { get; set; }
+    public decimal LastReading { get; set; }
 
     [BsonElement("lastReadOn")]
     public DateTime LastReadOn { get; set; }
